+++ {{cookiecutter.project_slug}}/.gitignore
@@ -1,2 +1,3 @@
 src/__pycache__
+src/.cache
 .vscode
