+++ {{cookiecutter.project_slug}}/src/api.py
@@ -3,80 +3,64 @@
     Below is a simple, incomplete but running example of an API.
 """
 
-import werkzeug
-from flask import Flask, abort
-from flask_restplus import Api, Resource, fields, reqparse
+from typing import List
+from fastapi import FastAPI, Query, HTTPException, File
+from pydantic import BaseModel, Schema
+
 import model_utils
 import utils
 
-app = Flask(__name__)
-
-api = Api(app, version='1.0', title='{{cookiecutter.project_name}}',
-          description="{{cookiecutter.project_short_description}}",
-          contact="")
-
-# We define api models below to specify the response for some routes
-TFL_MODEL = api.model("TFLite Model", {
-    "model_name": fields.String(required=True, description="The model's name"),
-    "inference_type": fields.String(required=True, description="The functionality of the model (e.g. image classification, object detection)")
-})
-
-PREDICTION_MODEL = api.model("Classification Result", {
-    "class_name": fields.String(required=True, description="The predicted class"),
-    "confidence": fields.Float(required=True, description="The classification score")
-})
-
-VERSION_MODEL = api.model("API Version", {
-    "version": fields.String(required=True, description="The API version")
-})
+app = FastAPI(version='1.0', title='{{cookiecutter.project_name}}',
+              description="{{cookiecutter.project_short_description}}")
 
 
-@api.route('/models')
-class Models(Resource):
-    # The .marshal_with and .marshal_list_with decorators
-    # are for describing the content of the request's response
-    @api.marshal_list_with(TFL_MODEL)
-    def get(self):
-        """ Returns the list of uploaded models """
-        try:
-            models = utils.get_models_info()
-        except FileNotFoundError:
-            abort(500, "The /models directory was not found!")
-
-        return models
+# We define api models below to document expected request and response bodies
+class VersionResponse(BaseModel):
+    # Schema is used to give more information on each attribute of the model
+    # use ellipsis ('...') to indicate the field is required
+    version: str = Schema(..., description="API version")
 
 
-@api.route('/version')
-class Version(Resource):
-    @api.marshal_with(VERSION_MODEL)
-    def get(self):
-        """ Returns the API version"""
-        return {"version": api.version}
+class ModelsResponse(BaseModel):
+    model_name: str
+    inference_type: str = Schema(..., description="Functionality of the model")
 
 
-# Below is a parser for parsing the arguments that the /predict route requires in HTTP requests
-prediction_parser = reqparse.RequestParser()
-prediction_parser.add_argument('image', type=werkzeug.datastructures.FileStorage, location='files',
-                               required=True, help="The image to perform inference on")
-prediction_parser.add_argument('model_name', type=str, required=True,
-                               help="The name of the model to be used for predictino")
+class PredictionResponse(BaseModel):
+    class_name: str = Schema(..., description="Predicted class")
+    confidence: float = Schema(..., description="Classification score")
 
 
-@api.route('/predict')
-@api.expect(prediction_parser)
-class Predict(Resource):
-    @api.marshal_with(PREDICTION_MODEL)
-    def post(self):
-        """ Runs the model loaded in memory to infer on a given image """
-
-        args = prediction_parser.parse_args(strict=True)
-        image = args['image']
-        model_name = args['model_name']
-
-        prediction = model_utils.predict(image, model_name)
-
-        return prediction
+# Below we define the routes for the API
+@app.get('/version', response_model=VersionResponse)
+async def get_version():
+    """ Returns the API version"""
+    return {"version": app.version}
 
 
-if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=5000, debug=True)
+@app.get("/models", response_model=List[ModelsResponse])
+async def get_models():
+    """ Returns the list of uploaded models """
+    try:
+        models = utils.get_models_info()
+    except FileNotFoundError:
+        raise HTTPException(500, detail="The /models directory was not found!")
+
+    return models
+
+
+@app.post('/predict', response_model=PredictionResponse)
+async def predict(
+    model_name: str = Query(..., description="Name of model for inference"),
+    image: bytes = File(..., description="Image to perform inference on"),
+):
+    """ Runs the model loaded in memory to infer on a given image """
+    try:
+        image = utils.convert_bytes_to_image(image)
+    except Exception as ex:
+        raise HTTPException(422, detail="Error while reading request image.\
+            Please make sure it is a valid image. " + str(ex))
+
+    prediction = model_utils.predict(image, model_name)
+
+    return prediction
