--- {{cookiecutter.project_slug}}/tests/test_{{cookiecutter.project_slug}}.py
@@ -1,1 +1,0 @@
-""" Contains unit tests """
